feat(ehr): remove uselses passthrough
Ref: ENG-139

Ref: #1040
